--- src/GameOver.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const GameOver = (props) => {
-    return (
-        <div className="game-area-info">
-            <h1>Game is over! Press <em>SPACE</em> to start.</h1>
-        </div>
-    )
-}
-export default GameOver;
