--- src/Pause.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Pause = () => {
-    return (
-        <h2 className="game-area-info">
-            Game Paused press <em>ESC</em> button to resume game
-        </h2>
-    )
-}
-
-export default Pause
