+++ src/components/GameComponents/Pause.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Pause = () => {
+    return (
+        <h2 className="game-area-info">
+            Game Paused press <em>ESC</em> button to resume game!
+            <br/>
+            You can change game setting by the menu below ↓
+        </h2>
+    )
+}
+
+export default Pause
